+++ src/display_queue.cpp
@@ -21,7 +21,7 @@
 //	SOFTWARE.
 //
 
-#include "display_queue.h"
+#include "libexecstream/display_queue.h"
 
 namespace daw {
 	display_queue::q_item::q_item( stream_type_t type, std::string item ) noexcept:
